requirements.txt
numpy==1.22.3
python-dateutil==2.8.2
requests==2.27.1
selenium==4.1.3
webdriver-manager==3.5.4
python-dotenv==0.20.0
pandas==1.4.1
XlsxWriter==3.0.3
opencv-python==4.5.5.64
